Cleaned up and setup Github pages
Cleaned up a lot of the extra files and stuff in the project. Moved all CSS info into one document and removed a lot of CSS as well and put it in "in line" styling instead. Created a folder for all icons. We also set up pages and will be created an index.html file to be the landing page for mobile at the moment