swapped index and land page. created home page
We turned the land page into the index page so that is now the default page. The index page is now the stopwatch page. We also created a home page with a link to the stopwatch page on it. Now that this is done: Full website navigation is possible from the login screen. Read the README.txt for more information on this