Started making maps - comments
Added the JS to create a map to the mobile JS file. Also added an area in the stopwatch file for the map to be just temporally. Also added more in-depth and helpful comments to all new account related pages and stopwatch. Will be going back in soon to add comments to other files