map is *almost* working
The map has now been added and integrated into the stopwatch. The map creation with data pushed into the array is working. Our issue is the location success function never runs. working on that atm. Updated other files to work on the internet instead of just locally. At the moment, there are NO bugs with currently made webpages except for the one map bug!