Implemented the $_SESSION ID
Created a new php file called verifyLoggedIn which is included on the top of any page that requires the user to be logged in for. This file automatically checks if the user is logged in. If not, it brings the user to the index page. Implemented this file into all created file that need this.